bash: fall back to flag string if `shellDryRun` is unavailable
`pkgs.stdenv.shellDryRun` is only available in the unstable channel at
the moment. While normally it would be feasible to make a package change
available to stable channels by way of an overlay, overriding `stdenv`
will cause a cascade of rebuilds since every derivation that depends on
it will no longer be able to pull a cached binary.

Despite the fallback, this change retains support for `extglob` by
copying the value of `shellDryRun` from its definition in nixpkgs.

https://github.com/NixOS/nixpkgs/blob/b20f90e94001f58571067ac1c49b71d1eb73bd52/pkgs/stdenv/generic/default.nix#L175